Cleanup of old methods
